requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
distlib==0.3.4
Django==3.2.13
djangorestframework==3.12.4
filelock==3.4.0
heroku==0.1.4
idna==3.3
platformdirs==2.4.0
psycopg2==2.9.3
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.10.0
